Align template genres and split type guards
